test: github actions diff of single test func
